feat: :sparkles: total -> home 이동 및 homepage의 새로운 쿼리 추가
- #103
