+++ app/src/page/home/home.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { CursusUserModule } from 'src/api/cursusUser/cursusUser.module';
+import { QuestsUserModule } from 'src/api/questsUser/questsUser.module';
 import { ScaleTeamModule } from 'src/api/scaleTeam/scaleTeam.module';
-import { ScaleTeamService } from 'src/api/scaleTeam/scaleTeam.service';
+import { ScoreModule } from 'src/api/score/score.module';
 import { HomeResolver } from './home.resolver';
 import { HomeService } from './home.service';
 
 @Module({
-  imports: [ScaleTeamModule, CursusUserModule],
-  providers: [HomeResolver, HomeService, ScaleTeamService],
+  imports: [ScaleTeamModule, CursusUserModule, QuestsUserModule, ScoreModule],
+  providers: [HomeResolver, HomeService],
 })
 export class HomeModule {}
